// src/modules/auth/constants/auth-message.enum.ts
export enum AuthMessage {
  LOGGED_IN = "Logged In OK. hi 👋",
  LOGGIN_FAILED = "Failed to loggin",
  LOGGED_OUT = "Logged out OK. bye 🙋‍♂️",
  REFRESH_TOKEN_NOT_FOUND = "Refresh token not found",
  TOKEN_NOT_FOUND = "Token not found",
  REFRESHED_TOKEN_OK = "Refreshed token OK. Welcome back 🙋‍♀️",
  SIGNATURE_INVALID = `User's signature verification failed`,
  ADDRESS_BANNED = "Address banned",
  INVALID_PHONE = "Số điện thoại không hợp lệ",
  CHECK_USE_PHONE = "Số điện thoại đã được sử dụng",
  NOT_FOUND = "Số điện thoại mật khẩu không chính xác",
  SEND_SMS = "Gửi mã code thành công",
  VERIFY_SMS = "Xác minh mã code thành công",
  UPDATE_USER = "Cập nhật thông tin thành công",
  ERR_500 = "Lỗi hệ thống",
  UPDATE_FIRST_ERR = "Không thể câo nhật thông tin",
  NOT_ACCESS = "Không có quyền truy cập",
  UPDATE_SUCCESS = "Cập nhật thông tin thành công",
  NOT_MATCH = "Mật khẩu và xác nhận mật khẩu không khớp",
  UPDATE_PASSWORD = "Cập nhật mật khẩu thành công",
  NOT_ACTIVE = "Tài khoản dã bị khoá",
  DELETE_SUCCESS = "Xóa tài khoản thành công",
}